Create Miami GP
Chef has finally got the chance of his lifetime to drive in the 
𝐹
1
F1 tournament. But, there is one problem. Chef did not know about the 107% rule and now he is worried whether he will be allowed to race in the main event or not.

Given the fastest finish time as 
𝑋
X seconds and Chef's finish time as 
𝑌
Y seconds, determine whether Chef will be allowed to race in the main event or not.

Note that, Chef will only be allowed to race if his finish time is within 107% of the fastest finish time.