Create Pass or Fail
Chef is struggling to pass a certain college course.

The test has a total of 
𝑁
N questions, each question carries 
3
3 marks for a correct answer and 
−
1
−1 for an incorrect answer. Chef is a risk-averse person so he decided to attempt all the questions. It is known that Chef got 
𝑋
X questions correct and the rest of them incorrect. For Chef to pass the course he must score at least 
𝑃
P marks.

Will Chef be able to pass the exam or not?